refactor: 오타로 인한 구조변경
